backend/pyproject.toml
[project]
name = "wishlists_backend"
version = "0.1.0"
description = "FastAPI backend for Wish Lists web app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi[standard]>=0.115.4",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.10.0",
    "sqlmodel>=0.0.22",
]
